+++ CHANGELOG.md
@@ -13,7 +13,7 @@ to [sourceforge feature requests](https://sourceforge.net/p/jabref/features/) by
 ## [Unreleased]
 
 ### Changed
- - Add support for proxy authentication via VM args, this implements Feature Request #388
+ - Add support for proxy authentication via VM args and GUI settings, this implements Feature Request #388
 
 ### Fixed